services/common/model/pyproject.toml
[tool.poetry]
name = "core"
version = "1.0.0"
description = "Model defined by SQLAlchemy and Alembic"
authors = ["thor <thor@wolpert.ca>"]
license = "BSD-3-Clause"
readme = "README.md"
packages = [{include = "core", from = "src"},
            {include = "migrations", from = "src"}
           ]

[tool.poetry.dependencies]
python = "^3.11"
Flask-SQLAlchemy = "^3.0.5"
pycountry = "^22.3.5"
datedelta = "^1.4"
geoalchemy2 = "^0.16.0"

sql-versioning = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/sql-versioning", branch = "main" }
nanoid = "^2.0.0"
sqlalchemy-utils = "^0.41.2"

[tool.poetry.group.dev.dependencies]
Flask = "^2.3.2"
Flask-Migrate = "^4.0.4"
flask-babel = "^3.1.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
black = "^23.12.1"
isort = "^5.13.2"
pg8000 = "^1.29.8"
freezegun = "^1.2.2"

[tool.black]
target-version =  ["py310", "py311", "py312"]
line-length = 120
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  migrations
  | devops
)/
'''

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true
skip_glob = ["migrations", "devops"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"